Add Tig Ol' Bitties V2.5
The left breast now copy the colors of the right indefinitely, the nipples and vag also maintains their indented colors indefinitely. as of writing this i have moved the plan of physics from v3 to v4, for i plan to add male genitalia and nipples placement depending on gender, both by v3.